Renaming folder and adjust file's  paths
