+++ src/renderer/franchiseRequestsRefuse/index.js
@@ -1,0 +1,2 @@
+import FranchiseRequestRefuse from "./FranchiseRequestRefuseScreen";
+export default FranchiseRequestRefuse;
