--- src/renderer/franchiseRequestsRefused/index.js
@@ -1,2 +1,0 @@
-import FranchiseRequestRefuse from ".//FranchiseRequestRefuseScreen";
-export default FranchiseRequestRefuse;
